requirements.txt
Flask==0.12.2
google-cloud-storage==1.4.0
gunicorn==19.7.1
numpy==1.13.3
scikit-learn==0.19.0
scipy==1.0.0